Uploaded MoM_bin.py
File has algorithm to calculate acceptance function for a q2 bin in any angular variable from the 4D full acceptance function. Needs to use the acceptance_mc file with all filters EXCEPT q2. Can find acceptance functions in costhetal, costhetak and phi for any bin in the binning scheme. Care must be taken with the acceptance in phi as it is calculated using a rescaling to the range -1 to 1, appropaite procedure for generating acceptance functions described in the docstrings.